udia v0.1.1 Udia.Web.Router.Helpers

Module with named helpers generated from Udia.Web.Router.

Functions

category_path(conn_or_endpoint, action)

category_path(conn_or_endpoint, action, params)

category_path(conn_or_endpoint, action, id, params)

category_url(conn_or_endpoint, action)

category_url(conn_or_endpoint, action, params)

category_url(conn_or_endpoint, action, id, params)

comment_path(conn_or_endpoint, action)

comment_path(conn_or_endpoint, action, params)

comment_path(conn_or_endpoint, action, id, params)

comment_url(conn_or_endpoint, action)

comment_url(conn_or_endpoint, action, params)

comment_url(conn_or_endpoint, action, id, params)

letsencrypt_path(conn_or_endpoint, action, id)

letsencrypt_path(conn_or_endpoint, action, id, params)

letsencrypt_url(conn_or_endpoint, action, id)

letsencrypt_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

post_path(conn_or_endpoint, action)

post_path(conn_or_endpoint, action, id)

post_path(conn_or_endpoint, action, id, params)

post_url(conn_or_endpoint, action)

post_url(conn_or_endpoint, action, id)

post_url(conn_or_endpoint, action, id, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_path(conn_or_endpoint, action, id, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)